Crate statime

Statime is a library providing an implementation of PTP version 2.1 (IEEE1588-2019). It provides all the building blocks to setup PTP ordinary and boundary clocks.

Note: We are currently planning a major overhaul of the library. This will also result in significant changes to the public API.

Device interfaces

statime is designed to be able to work with many different underlying platforms, including embedded targets. This does mean that it cannot use the standard library and platform specific libraries to interact with the system clock and to access the network. That needs to be provided by the user of the library.

The statime crate defines a Clock interface that provide access to the system clock. The [NetworkRuntime] and [NetworkPort] abstractions provide the needed glue to access the network.

On modern linux kernels, the statime-linux crate provides ready to use implementations of these interfaces. For other platforms the user will need to implement these themselves.

Clock identities

All ptp clocks in a network need a unique clock identity. One way to achieve this is to use (one of) the device’s mac address to generate this identifier. As this requires platform specific code to get the mac address, this library does not implement this. Rather, direct access is given to the clock identity type, and the user can create one from a mac address by storing it in the first six bytes of the clock identifier, setting the remaining bytes to 0. For more details on the exact specification of the generation procedure, see IEEE1588-2019 section 7.5.2.2.2

Structs

BasicFilter

A simple averaging filter

ClockIdentity

The identity of a PTP node.

ClockQuality

A description of the accuracy and type of a clock.

Duration

A duration is a span of time that can also be negative.

FuzzMessage

InBmca

InstanceConfig

Interval

Measurement

A single measurement as produced by a PTP port.

Port

A single port of the PTP instance

PortActionIterator

Guarantees to end user: Any set of actions will only ever contain a single time critical send

PortConfig

Configuration items of the PTP PortDS dataset. Dynamical fields are kept as part of crate::port::Port.

PtpInstance

A PTP node.

Running

SdoId

A wrapper type for PTP Sdo Identifiers.

Time

Time represents a specific moment in time.

TimePropertiesDS

A concrete implementation of the PTP Time Properties dataset (IEEE1588-2019 section 8.2.4

TimestampContext

Enums

ClockAccuracy

How accurate the underlying clock device is expected to be when not synchronized.

DelayMechanism

Which delay mechanism a port is using.

LeapIndicator

PortAction

TimeSource

What the time values for a system are derived from

Constants

MAX_DATA_LEN

Traits

Clock

Clock manipulation and querying interface

Filter

A filter for post-processing time measurements.