Trait statime::Filter

pub trait Filter {
    type Config: Clone;

    // Required methods
    fn new(config: Self::Config) -> Self;
    fn measurement<C: Clock>(
        &mut self,
        m: Measurement,
        clock: &mut C
    ) -> FilterUpdate;
    fn delay(&mut self, delay: Duration) -> Duration;
    fn update<C: Clock>(&mut self, clock: &mut C) -> FilterUpdate;
    fn demobilize<C: Clock>(self, clock: &mut C);
}

A filter for post-processing time measurements.

Filters are responsible for dealing with the network noise, and should average out the input a bit so minor network variations are not immediately reflected in the synchronization of the clock.

This crate provides a simple BasicFilter which is suitable for most needs, but users can implement their own if desired.

Required Associated Types

type Config: Clone

Required Methods

fn new(config: Self::Config) -> Self

Create a new instance of the filter.

fn measurement<C: Clock>( &mut self, m: Measurement, clock: &mut C ) -> FilterUpdate

Put a new measurement in the filter. The filter can then use this to adjust the clock

fn delay(&mut self, delay: Duration) -> Duration

Handle a new measurement of the delay to the master.

fn update<C: Clock>(&mut self, clock: &mut C) -> FilterUpdate

Update initiated through [FilterUpdate::next_update] timeout.

fn demobilize<C: Clock>(self, clock: &mut C)

Handle ending of time synchronization from the source associated with this filter.